--- SEL4C/SEL4C/urls.py
@@ -1,14 +1,0 @@
-from django.urls import include, path
-from rest_framework import routers
-from app1 import views
-
-router = routers.DefaultRouter()
-router.register(r'users', views.UserViewSet)
-router.register(r'groups', views.GroupViewSet)
-
-#Wire up our API using automatic URL lookup.
-#additionally, we include login URLs for browsable API.
-urlpatterns = [
-    path('', include(router.urls)),
-    path('api-auth/', include('rest_framework.urls', namespace='rest_framework'))
-]
