--- SEL4C/app1/serializers.py
@@ -1,12 +1,0 @@
-from django.contrib.auth.models import User, Group
-from rest_framework import serializers
-
-class UserSerializer(serializers.HyperlinkedModelSerializer):
-    class Meta:
-        model = User
-        fields = ['url', 'username', 'email', 'groups']
-
-class GroupSerializer(serializers.HyperlinkedModelSerializer):
-    class Meta:
-        model = Group
-        fields = ['url', 'name']
