+++ TC2007B_IntSegInfRedSisSof_BE-451-A01662274_IVA/SEL4C/SEL4C/urls.py
@@ -1,0 +1,19 @@
+from django.urls import include, path
+from rest_framework import routers
+from app1 import views
+from drf_spectacular.views import SpectacularAPIView, SpectacularSwaggerView, SpectacularRedocView
+
+router = routers.DefaultRouter()
+router.register(r'users', views.UserViewSet)
+router.register(r'groups', views.GroupViewSet)
+router.register(r'home',views.HomeViewSet)
+
+#Wire up our API using automatic URL lookup.
+#additionally, we include login URLs for browsable API.
+urlpatterns = [
+    path('', include(router.urls)),
+    path('api-auth/', include('rest_framework.urls', namespace='rest_framework')),
+    path('api/schema/', SpectacularAPIView.as_view(), name='schema'),
+    path('api/schema/swagger/',SpectacularSwaggerView.as_view(url_name='schema'),name='swagger-ui'),
+    path('api/schema/redoc/',SpectacularRedocView.as_view(url_name='schema'),name='redoc')
+]
