+++ TC2007B_IntSegInfRedSisSof_BE-451-A01662274_IVA/SEL4C/SEL4C/wsgi.py
@@ -1,0 +1,16 @@
+"""
+WSGI config for SEL4C project.
+
+It exposes the WSGI callable as a module-level variable named ``application``.
+
+For more information on this file, see
+https://docs.djangoproject.com/en/4.2/howto/deployment/wsgi/
+"""
+
+import os
+
+from django.core.wsgi import get_wsgi_application
+
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'SEL4C.settings')
+
+application = get_wsgi_application()
