+++ TC2007B_IntSegInfRedSisSof_BE-451-A01662274_IVA/SEL4C/app1/admin.py
@@ -1,0 +1,3 @@
+from django.contrib import admin
+
+# Register your models here.
