+++ TC2007B_IntSegInfRedSisSof_BE-451-A01662274_IVA/SEL4C/app1/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class App1Config(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'app1'
+    #name = 'SEL4C.app1'
