+++ TC2007B_IntSegInfRedSisSof_BE-451-A01662274_IVA/SEL4C/app1/models.py
@@ -1,0 +1,11 @@
+from django.db import models
+
+class HomeModel(models.Model):
+    title = models.CharField(max_length=100)
+    content = models.TextField()
+    created_at = models.DateTimeField(auto_now_add=True)
+
+    def __str__(self):
+        return self.title
+    class Meta:
+        app_label = 'app1'
