+++ TC2007B_IntSegInfRedSisSof_BE-451-A01662274_IVA/SEL4C/app1/tests.py
@@ -1,0 +1,3 @@
+from django.test import TestCase
+
+# Create your tests here.
