+++ TC2007B_IntSegInfRedSisSof_BE-451-A01662274_IVA/SEL4C/app1/views.py
@@ -1,0 +1,30 @@
+from django.contrib.auth.models import User, Group
+from rest_framework import viewsets
+from rest_framework import permissions
+from app1.serializers import UserSerializer, GroupSerializer
+from .models import HomeModel
+from .serializers import HomeSerializer
+
+class UserViewSet(viewsets.ModelViewSet):
+    """
+    API endpoint that allows users to be viewed or edited.
+    """
+    queryset = User.objects.all().order_by('-date_joined')
+    serializer_class = UserSerializer
+    permission_classes = [permissions.IsAuthenticated]
+
+class GroupViewSet(viewsets.ModelViewSet):
+    """
+    API endpoint that allows groups to be viewed or edited
+    """
+    queryset = Group.objects.all()
+    serializer_class = GroupSerializer
+    permission_classes = [permissions.IsAuthenticated]
+
+class HomeViewSet(viewsets.ModelViewSet):
+    '''
+    API endpoint that allows MyModel to be viewed or edited.
+    '''
+    queryset = HomeModel.objects.all() #Set the queryset for the view
+    serializer_class = HomeSerializer #Set the serializer class
+    permission_classes = [permissions.IsAuthenticated] #Set the permission classes
